+++ src/onegov.onboarding/onegov/onboarding/tests/conftest.py
@@ -1,9 +1,9 @@
+import morepath
 import onegov.core
 import onegov.onboarding
 import os.path
 import pytest
 
-from morepath import setup
 from onegov.core.utils import scan_morepath_modules
 from uuid import uuid4
 
@@ -11,9 +11,8 @@ from uuid import uuid4
 @pytest.yield_fixture(scope="function")
 def onboarding_app(postgres_dsn, temporary_directory, smtp):
 
-    config = setup()
-    scan_morepath_modules(onegov.onboarding.OnboardingApp, config)
-    config.commit()
+    scan_morepath_modules(onegov.onboarding.OnboardingApp)
+    morepath.commit(onegov.onboarding.OnboardingApp)
 
     app = onegov.onboarding.OnboardingApp()
     app.namespace = 'test_' + uuid4().hex